refactor: Core 패지키 구조로 변경
