--- Projects/Core/AppCore/README.md
@@ -1,3 +1,0 @@
-# Core
-
-A description of this package.
