--- Projects/Core/AppCore/Sources/AppCore/AppCore.swift
@@ -1,1 +1,0 @@
-public struct AppCore {}
