--- Projects/Core/AppCore/Tests/AppCoreTests/AppCoreTests.swift
@@ -1,4 +1,0 @@
-@testable import AppCore
-import XCTest
-
-final class AppCoreTests: XCTestCase {}
