--- Projects/Core/AppFoundation/Package.resolved
@@ -1,23 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "ribs",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/uber/RIBs.git",
-      "state" : {
-        "branch" : "main",
-        "revision" : "0b0948a68f4836f504a09529176e8da90909f846"
-      }
-    },
-    {
-      "identity" : "rxswift",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/ReactiveX/RxSwift",
-      "state" : {
-        "revision" : "b4307ba0b6425c0ba4178e138799946c3da594f8",
-        "version" : "6.5.0"
-      }
-    }
-  ],
-  "version" : 2
-}
