--- Projects/Core/AppFoundation/README.md
@@ -1,3 +1,0 @@
-# AppFoundation
-
-A description of this package.
