--- Projects/Core/AppFoundation/Tests/AppFoundationTests/AppFoundationTests.swift
@@ -1,4 +1,0 @@
-@testable import AppFoundation
-import XCTest
-
-final class AppFoundationTests: XCTestCase {}
