--- Projects/Core/Entity/README.md
@@ -1,3 +1,0 @@
-# Entity
-
-A description of this package.
