--- Projects/Core/Entity/Tests/EntityTests/EntityTests.swift
@@ -1,4 +1,0 @@
-@testable import Entity
-import XCTest
-
-final class EntityTests: XCTestCase {}
