--- Projects/Core/Network/README.md
@@ -1,3 +1,0 @@
-# Network
-
-A description of this package.
