+++ Projects/Core/Package.resolved
