--- Projects/Core/Platform/README.md
@@ -1,3 +1,0 @@
-# Platform
-
-A description of this package.
