+++ Projects/Core/Sources/Network/Plugins/NetworkErrorDescriptor.swift
