+++ Projects/Core/Sources/Platform/LocalDataSupplier.swift
