+++ Projects/Core/Sources/Platform/Protocols/AnyOptional.swift
